entrez.py: Change arguments to on_search()
The signature of on_search() is now:
  on_search(term, db, tool, dbfrom=None, **params)
with a different semantics than before (db is the db where the
tool is applied, and dbfrom is the db where the search is performed).

This is similar to the fields in the elink tool.

There are also some cosmetic changes (using "yield from").

Thanks Adrian Lopez for the suggestion!